requirements.txt
setuptools~=57.4.0
pyserial~=3.5